+++ R/zzz.R
@@ -1,5 +1,9 @@
-globalVariables(c("code", "label_en", "new_code", "broader", "external",
-                  "extCode", ".", "sourceID", "sourceName", "description",
+globalVariables(c("id", "label_en", "new_id", "has_broader", "external",
+                  "extCode", ".", "source_id", "source_label", "description",
                   "homepage", "notes", "level", "undef", "topID", "top2D",
-                  "nestedID", "nextID", "newClass"
+                  "nestedID", "nextID", "newClass", "extid", "broader",
+                  "class_label", "external_id", "has_narrower_match",
+                  "has_close_match", "has_exact_match", "has_broader_match",
+                  "new_exact_match", "new_narrower_match", "new_close_match",
+                  "new_broader_match", "avail", "has_source", "label", "newid"
                   ))