+++ NEWS.md
@@ -1,3 +1,14 @@
+# ontologics 0.4.0
+
+- adapt the onto-class structure according to issue https://github.com/luckinet/ontologics/issues/6; this entails that concepts and classes are summarised into their respective slot, with sub-tables "harmonised" and "external", featuring specific columns.
+- change all functions to support this new structure
+- this changes the behaviour of the function `new_concept()` insofar that this only creates harmonised concepts (so no source needs to be provided here anymore) and of the funciton `new_mapping()` insofar that this now creates external concepts and their mapping to harmonised concepts (and thus writes into the sub-table "external" and into the "mapping-columns" of the sub-table "harmonised")
+
+# ontologics 0.3.2
+
+- adapt column names according to issue https://github.com/luckinet/ontologics/issues/5
+- revise how to extract concents via `get_concept()`. Now you have to provide a table that contains the column name on which to subset and the values in that column that should be filtered from the ontology.
+
 # ontologics 0.3.1
 
 - adapt `new_concept()` so that also concepts can be defined that have no class (for various reasons). This now gives a warning, and inserts the concepts with an "undefined" class.